Rename `reverse-matches.ts` to `scope-limiter.ts` and `isReverseMatch` to `isInScope`
