+++ hooks/cascading-scans/hook/Dockerfile
@@ -15,10 +15,10 @@ RUN mkdir -p /home/app
 WORKDIR /home/app
 COPY package.json package-lock.json ./
 RUN npm ci
-COPY hook.ts scan-helpers.ts reverse-matches.ts kubernetes-label-selector.ts ./
+COPY hook.ts scan-helpers.ts scope-limiter.ts kubernetes-label-selector.ts ./
 RUN npm run build
 
 FROM ${namespace:-securecodebox}/hook-sdk-nodejs:${baseImageTag:-latest}
 WORKDIR /home/app/hook-wrapper/hook/
 COPY --from=install --chown=app:app /home/app/node_modules/ ./node_modules/
-COPY --from=build --chown=app:app /home/app/hook.js /home/app/hook.js.map /home/app/scan-helpers.js /home/app/scan-helpers.js.map /home/app/reverse-matches.js /home/app/reverse-matches.js.map /home/app/kubernetes-label-selector.js /home/app/kubernetes-label-selector.js.map ./
+COPY --from=build --chown=app:app /home/app/hook.js /home/app/hook.js.map /home/app/scan-helpers.js /home/app/scan-helpers.js.map /home/app/scope-limiter.js /home/app/scope-limiter.js.map /home/app/kubernetes-label-selector.js /home/app/kubernetes-label-selector.js.map ./
